Added mixed reward function for stabilizers and Pauli.
Added attribute '_epidode_return' to the environment to calculate
undiscouted return up to the current time step. Also added the new
reward function which calls 'reward_stabilizers' at intermediate
times and 'reward_pauli' on the last time step. The last reward is
scaled by the episode return. Not sure if this is a good reward fn,
will be testing it in the following days.