Added tensorflow_benchmarking.py script
This script is a cleaned up version of the previous script
tensorflow_vs_qutip. There is a way to improve it even more
when the next stable release of tf-gpu comes out. Currently
tf.linalg.diag doesn't work correctly in tf2.1, but with this
function there is a way to generate annihilation ops better.