Switched to using dictionary actions instead of stacked tensors.
Now the action that environment accepts is a dictionary with keys
'alpha', 'beta' etc. This allowed to significantly simplify a lot
of code and make it more general. Now it will be very easy to add
more action components without writing any specialized wrappers etc.
Some of the major changes that go with this commit are:
1) gkp environment class is updated and simplified
2) action wrapper is now unified and works with phase-estimation
and Baptiste-style protocols.
3) ScriptedPolicy class is also unified in the same way. No need
to have separate classes for different types of quantum circuit.
4) All action scripts are slightly modified: order in which alpha
script is presented is shifted by 1. This was done to simplify the
ScriptedPolicy.
5) action_histogram_feedback_trim is modified and now will work the
same for policies that only learn feedback. Therefore, in the next
commit I will delete action_histogram_feedback.
5) Other analysis scripts such as T1 measurement are slightly up-
dated to match new structure of actions.