Added option for polar representation of complex amplitudes.
The representation is specified by the flag 'complex_form' which
controls how the environment interprets the 2D vectors for ampli
tudes. Also ActionWrapper, train_eval and plotting scripts are
slightly modified to be compatible with this new option. All pre
vious trainined policies are still compatible if the flag is set
to 'cartesian'.