Pauli reward can be used with quantum_circuit_v2 and few more fixes.
It's a quick and dirty implementation so needs to be cleaned up later.
Other fixes include the correct mask which allows to learn which stabi
lizer to measure and which pauli to apply after trimming. This is done
by interpreting the agents output as the quadrant in which the displac
ement needs to be done, but disregarding the exact magnitude that the
agent has predicted. Also, in PPO moved fc_layers from input of RNN to
output.