Update benchmarking framework for D(alpha) and add error analysis
We're interested in comparing the performance of different algorithms
for generation of displacement operators along with various approaches
to parallelization and performance. The displacement operators are
needed to simulate the environment during the reinforcement learning
process.

We also add a number of different error analyses that were evaluated
throughout the benchmarking process. The native expm method is also
approximating, so it's not necessarily an ideal reference. The final
approach chosen is to compare the coefficients of coherent states
(generated with the operators) against the analytic formula.

We also refactor the benchmarking framework a bit to ease the
introduction of new algorithms and to increase code reuse and
readability.

Co-authored-by: Vladimir Sivak <sivak.volodymyr@gmail.com>