Modified action wrapper and action scripts to allow easy masking.
This is likely to go away in the future, but for now since I am actively
experimenting with using a mask, it's better to clean it up a bit. There
are too many rings and bells now: actions can be masked or not masked. If
they are maksed, they can be scripted or projected from the quadrant which
is learned. And it's not the same for 'alpha' and 'beta' because 'alpha'
can't be scripted, so there is only a projection from qudrant option for it.
All this makes my action wrapper look ugly and I can't wait to use some
Okkam razor to nicely delete all these options.