Implement Baker-Campbell-Hausdorff version of matrix exponential
We implement the BCH formula to calculate the displacement operator. We
use the q, p operators and diagonalize them in advance. The resulting
diagonal matrices can be exponentiated directly along the diagonal.

There is some subtlety due to our truncated Fock space, as [a_dag, a] is
no longer -I. The highest value of the numerical commutator is seen as:
https://quantumcomputing.stackexchange.com/questions/9779/commutation-of-bosonic-operators-on-finite-hilbert-space

However, the improvement in speed and reduction in memory (compared to
the extra Pade approximants in the TF expm implementation) allows us to
use complex128 instead of complex64. In comparing the error of coherent
state coefficients, the BCH approach is not only faster, but also more
accurate in the higher probability states.

On a NVIDIA V100-SXM2 GPU with a 1600x1600 Hilbert space, we get a ~100x
improvement for an entirely in-memory batch and a ~10x improvement for a
tf.data.Dataset batched run.

Co-authored-by: Vladimir Sivak <sivak.volodymyr@gmail.com>