Simplified action wrapper. Now it takes nested actions as input.
The actor neural net now will produce a dictionary of tensor actions
instead of one big tensor which is later sliced into pieces by the
wrapper. The wrapper will  simply add come action components to the
dictionary or decide if the scripted action should be used instead.
The observations still need to be flattened into a single tensor.