Implement a separate set of Simulator environment classes
We decouple dependencies between the GKP base class, the
environment-specific OscillatorGKP/OscilatorQubitGKP classes, and the
simulation itself.

There are now 3 separate areas of responsibility:
* The GKP base class is enforced as abstract, since it is intended to be
mixed with specific circuit implements of the protocol.
* The simulation-related pieces are pulled out into their own classes in
simulator, which can be used independently of the GKP inclusion.
* The OscillatorGKP and OscillatorQubitGKP classes now multiply-inherit
from a specific simulation definition, a batch operator mixin, and the
GKP base class. It implements the GKP protocol circuit.

Apart from this refactoring, the main "numerical" change to be careful
with is the removal of the explicit batch parameter from the simulation
in the operators defined on the space and the Kraus ops. Instead, we rely
on TF broadcasting semantics, replacing batch_dot with matvec for op *
psi, and with a @ b for operator composition.